Abstract the classes out into PubMed module.
